refactor(components): move `removePageCidMap()` to reduce cross-context coupling
• Move `removePageCidMap()` from `content/segmentSubmission.ts` into
  `utils/exporter.ts`, thereby preventing the components from importing
  content modules, which previously caused unnecessary code bloat during
  build. (dev build of Chrome extension size: ~75.4 MB → ~66.8 MB)

---
Ref: https://github.com/hanydd/BilibiliSponsorBlock/pull/281#discussion_r3178129496
