+++ src/components/SponsorTimeEditComponent.tsx
@@ -4,7 +4,6 @@ import * as React from "react";
 import * as CompileConfig from "../../config.json";
 import Config from "../config";
 import { keybindToString } from "../config/config";
-import { removePageCidMap } from "../content/segmentSubmission";
 import { ContentContainer } from "../ContentContainerTypes";
 import { showMessage } from "../render/MessageNotice";
 import { RectangleTooltip } from "../render/RectangleTooltip";
@@ -12,6 +11,7 @@ import { asyncRequestToServer } from "../requests/requests";
 import { ActionType, Category, ChannelIDStatus, SponsorTime } from "../types";
 import { DEFAULT_CATEGORY } from "../utils/categoryUtils";
 import { defaultPreviewTime } from "../utils/constants";
+import { removePageCidMap } from "../utils/exporter";
 import { getFormattedTime, getFormattedTimeToSeconds } from "../utils/formating";
 import { getFrameRate, getVideo } from "../utils/video";
 import { SelectorOption } from "./SelectorComponent";
